refactor: make showPageMask into a hook to remove the need to pass dispatch
Create a hook to expose `showPageMask` avoiding the need for a method
taking dispatch from outside.
