+++ apps/storefront/src/components/index.ts
@@ -11,7 +11,7 @@ export { default as B3Tip } from './B3Tip';
 export { default as CheckoutTip } from './extraTip/CheckoutTip';
 export { default as GlobalDialog } from './extraTip/GlobalDialog';
 export * from './HeadlessController';
-export { B3PageMask, Loading, showPageMask } from './loading';
+export { B3PageMask, Loading, usePageMask } from './loading';
 export { default as B3HoverButton } from './outSideComponents/B3HoverButton';
 export { default as B3MasquradeGobalTip } from './outSideComponents/B3MasquradeGobalTip';
 export { default as RegisteredCloseButton } from './RegisteredCloseButton';
