+++ apps/storefront/src/components/loading/index.ts
@@ -1,6 +1,6 @@
 import Loading from './Loading';
 
 export { default as B3PageMask } from './B3PageMask';
-export { default as showPageMask } from './B3showPageMask';
+export { usePageMask } from './B3showPageMask';
 
 export { Loading };
