+++ apps/storefront/src/pages/Dashboard/index.tsx
@@ -1,17 +1,16 @@
-import { useContext, useEffect, useRef, useState } from 'react';
+import { useEffect, useRef, useState } from 'react';
 import { useLocation } from 'react-router-dom';
 import { useB3Lang } from '@b3/lang';
 import MoreHorizIcon from '@mui/icons-material/MoreHoriz';
 import { Box, IconButton, Menu, MenuItem } from '@mui/material';
 import { styled } from '@mui/material/styles';
 
-import { showPageMask } from '@/components';
+import { usePageMask } from '@/components';
 import B3FilterSearch from '@/components/filter/B3FilterSearch';
 import B3Spin from '@/components/spin/B3Spin';
 import { B3PaginationTable } from '@/components/table/B3PaginationTable';
 import { TableColumnItem } from '@/components/table/B3Table';
 import { useSort } from '@/hooks';
-import { GlobaledContext } from '@/shared/global';
 import { superAdminCompanies } from '@/shared/service/b2b';
 import { useAppSelector } from '@/store';
 import { endMasquerade, startMasquerade } from '@/utils/masquerade';
@@ -106,7 +105,8 @@ function B3Mean({ isMasquerade, handleSelect, startActing, endActing }: B3MeanPr
 }
 
 function Dashboard(props: PageProps) {
-  const { dispatch } = useContext(GlobaledContext);
+  const showPageMask = usePageMask();
+
   const customerId = useAppSelector(({ company }) => company.customer.id);
   const b2bId = useAppSelector(({ company }) => company.customer.b2bId);
 
@@ -171,7 +171,7 @@ function Dashboard(props: PageProps) {
 
   const endActing = async () => {
     try {
-      showPageMask(dispatch, true);
+      showPageMask(true);
       if (typeof b2bId === 'number') {
         await endMasquerade({
           b2bId,
@@ -181,7 +181,7 @@ function Dashboard(props: PageProps) {
         ...filterData,
       });
     } finally {
-      showPageMask(dispatch, false);
+      showPageMask(false);
     }
   };
